mod: move prover type in client crate
