+++ crates/client/Cargo.toml
@@ -12,10 +12,10 @@ mockito = "0.31"
 ethrex-common = { workspace = true }
 ethrex-l2-common = { workspace = true }
 ethrex-rpc = { workspace = true }
+zkvm_interface = { workspace = true }
 
 # mojave
 mojave-signature = { workspace = true }
-mojave-chain-utils = { workspace = true }
 
 futures = { workspace = true }
 hex = { workspace = true }
