+++ crates/proof-coordinator/Cargo.toml
@@ -17,4 +17,4 @@ ethrex-storage-rollup = { workspace = true }
 ethrex-blockchain = { workspace = true }
 zkvm_interface = { workspace = true }
 
-mojave-chain-utils = { workspace = true }
+mojave-client = { workspace = true }
