+++ crates/prover/src/rpc/mod.rs
@@ -17,11 +17,11 @@ use ethrex_rpc::{
     utils::{RpcRequest, RpcRequestId},
 };
 
-use mojave_chain_utils::{
-    prover_types::{ProofResponse, ProofResult},
-    rpc::rpc_response,
+use mojave_chain_utils::rpc::rpc_response;
+use mojave_client::{
+    MojaveClient,
+    types::{ProofResponse, ProofResult},
 };
-use mojave_client::MojaveClient;
 
 pub struct ProverRpcContext {
     pub aligned_mode: bool,
