+++ crates/utils/src/lib.rs
@@ -2,6 +2,5 @@ pub mod initializer;
 pub mod logging;
 pub mod network;
 pub mod options;
-pub mod prover_types;
 pub mod rpc;
 pub mod unique_heap;
